refactor(tabs): nest Tab folder inside Tabs folder
