+++ src/components/navigation/Tabs/Tab/stories/Tab.stories.tsx
